compliance_auditor/requirements.txt
google-generativeai>=0.8.3
chromadb>=0.5.0
pypdf>=3.17.0
python-dotenv>=1.0.0
Pillow>=10.0.0
tqdm>=4.66.0
pandas>=2.0.0